server_flow/pyproject.toml
[tool.poetry]
name = "server-flow"
version = "0.1.0"
description = "Server Flow — custom server provider example with dotflow"
authors = ["FernandoCelmer <email@fernandocelmer.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
dotflow = {version = "0.15.0.dev6", allow-prereleases = true}
requests = "^2.32.3"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
